+++ src/taskAutom/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "7.15.4"
+__version__ = "7.16.2"
 __author__ = 'Lucas Aimaretto'